python/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.0.3
supabase>=2.3.0
python-dotenv>=1.0.0
rapidfuzz>=3.5.2
openpyxl>=3.1.2
